Extensions moved to config. Fileprop optimization. Dryrun mode added.

+++ mover.py
@@ -9,19 +9,21 @@ import fileprop
 
 class Mover(object):
     OUT_SUBDIR_CFG = {
-        fileprop.FileProp.IMAGE: 'out_subdir_image',
-        fileprop.FileProp.VIDEO: 'out_subdir_video',
-        fileprop.FileProp.AUDIO: 'out_subdir_audio',
+        fileprop.IMAGE: 'out_subdir_image',
+        fileprop.VIDEO: 'out_subdir_video',
+        fileprop.AUDIO: 'out_subdir_audio',
     }
 
-    def __init__(self, config, input_path, output_path, filenames):
+    def __init__(self, config, input_path, output_path, filenames, dryrun):
         self.__config = config
         self.__input_path = input_path
         self.__output_path = output_path
         self.__filenames = filenames
+        self.__dryrun = dryrun
         self.__move_mode = int(config['main']['move_mode'])
         self.__remove_garbage = int(config['main']['remove_garbage'])
         self.__stat = {'total': len(filenames)}
+        self.__file_prop = fileprop.FileProp(self.__config)
 
     def run(self):
         self.__stat['moved'] = 0
@@ -33,29 +35,30 @@ class Mover(object):
         res = []
         for fname in self.__filenames:
             try:
-                new_fname = self.__move_file(fname)
+                prop = self.__file_prop.get(fname)
+                new_fname = self.__move_file(fname, prop)
                 if new_fname:
-                    res.append(new_fname)
+                    res.append((fname, new_fname, prop))
             except Exception as ex:
                 logging.error('Move files exception: %s' % ex)
                 self.__stat['errors'] += 1
 
             self.__stat['processed'] += 1
+
         return res
 
-    def __move_file(self, fname):
-        prop = fileprop.FileProp(self.__config, fname)
-
-        if prop.type() == prop.GARBAGE:
+    def __move_file(self, fname, prop):
+        if prop.type() == fileprop.GARBAGE:
             if self.__remove_garbage:
-                os.remove(fname)
+                if not self.__dryrun:
+                    os.remove(fname)
                 logging.info('removed "%s"' % fname)
                 self.__stat['removed'] += 1
             else:
                 self.__stat['skipped'] += 1
             return None
 
-        if prop.type() == prop.OTHER or prop.time() is None:
+        if prop.type() == fileprop.IGNORE or prop.time() is None:
             self.__stat['skipped'] += 1
             return None
 
@@ -68,15 +71,19 @@ class Mover(object):
 
             path = os.path.join(self.__output_path, type_subdir, date_subdir)
             if not os.path.isdir(path):
-                os.makedirs(path)
+                if not self.__dryrun:
+                    os.makedirs(path)
+                logging.info('dir "%s" created' % path)
 
             fullname = prop.out_name_full(path)
             if self.__move_mode:
-                shutil.move(fname, fullname)
+                if not self.__dryrun:
+                    shutil.move(fname, fullname)
                 logging.info('"%s" moved "%s"' % (fname, fullname))
                 self.__stat['moved'] += 1
             else:
-                shutil.copy2(fname, fullname)
+                if not self.__dryrun:
+                    shutil.copy2(fname, fullname)
                 logging.info('"%s" copied "%s"' % (fname, fullname))
                 self.__stat['copied'] += 1
 
@@ -87,7 +94,8 @@ class Mover(object):
                 return None
             else:
                 new_fname = prop.out_name_full()
-                os.rename(fname, new_fname)
+                if not self.__dryrun:
+                    os.rename(fname, new_fname)
                 logging.info('"%s" renamed "%s"' % (fname, new_fname))
                 self.__stat['moved'] += 1
                 return new_fname
